requirements.txt
openai>=1.46.0
google-generativeai>=0.8.0
pillow
pyautogui
pytesseract
python-dotenv
mss
